Use tracepoints instead of kprobes for syscalls
This patch changes the way we attach to syscalls.
Instead of using a kprobe for each syscall, we use raw_syscalls
tracepoint as an entry point, and dispatch to the relevant syscall
handler using tail calls.
We first try to use the newer raw tracepoints (kernel ver >= 4.17), as
it is more performant, and fallback to "regular" tracepoint otherwise.

This change has the following advantages:
1. Stability between kernel versions
2. Better performance (see https://lwn.net/Articles/748352 for
   comparison)
3. Faster start time and stopping of tracee as there are less probes
   (only one for all syscalls)
4. It will allow us to support 32 bit userspace programs without adding
   additional probes