Add sched_process_exit event
Add sched_process_exit event and set it as essential.
Set do_exit event as not essential.
As tracepoints are considered a more stable API, and as there are
environments where tracepoints are enabled but kprobes are not,
it is preferable to use sched_process_exit and not do_exit as an
essential event.