add external package as a module
the external package was originally created for consuming tracee-ebpf events without having to import and build entire tracee-ebpf project, but since it was under the same module that defeated the purpose of a separate package. this commit is a preperation for a future commit which will import external package as a go module from this commit.