tracee-ebpf: sys_enter: use temp args_t struct pointer for no args
Fixes: #851

Somehow eBPF verifier did not like CO-RE object txt using the stack
address pointer from args_tmp as pointer for save_args() inline
function:

    libbpf: prog 'tracepoint__raw_syscalls__sys_enter': failed to
    attach to raw tracepoint 'sys_enter': Invalid argument

This commit creates a temporary args_t structure in the stack just
before calling save_args() and passing the structure pointer as an
argument. That is enough to allow the eBPF program to be loaded.