tracee-rules: Allow compiling and evaluating all Rego signatures at once
This patch adds new implementation of types.Signature that compiles
all Rego signatures into one Rego policy, which is then evaluated
with input events. This implementation is more efficient in terms
of number of memory allocations and number of objects created to
evaluate a Rego policy.

The new implementation is activated by specifying the --rego-aio
flag.

Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>