tracee-ebpf: add container cgroup ctime
This PR adds cgroup ctime information which can be used to track
containers start time.

Instead of using the WalkDir function to find a specific cgroup,
this PR implements a recursive function that is more efficient as it
only stats a directory if its name has a given suffix. By doing so,
performance is 5 times better than the use of WalkDir for this specific
case.