tracee-ebpf: fix kernfs_node CORE access in RHEL8
Fixes: #1171

RHEL8 4.18 kernels have a different kernfs_node struct. You can see the
name rh_kabi_hiddenXXX which is part of the RH KABI abstraction header
(an attempt of RH to keep KABI compatible between their back-ports):

struct kernfs_node {
        atomic_t count;
        atomic_t active;
        struct kernfs_node *parent;
        const char *name;
        struct rb_node rb;
        const void *ns;
        unsigned int hash;
        union {
                struct kernfs_elem_dir dir;
                struct kernfs_elem_symlink symlink;
                struct kernfs_elem_attr attr;
        };
        void *priv;
        union {
                u64 id;
                struct {
                        union kernfs_node_id id;
                } rh_kabi_hidden_172;
                union {         };
        };
        short unsigned int flags;
        umode_t mode;
        struct kernfs_iattrs *iattr;
};

while the regular v4.18 struct is:

struct kernfs_node {
        atomic_t count;
        atomic_t active;
        struct kernfs_node *parent;
        const char *name;
        struct rb_node rb;
        const void *ns;
        unsigned int hash;
        union {
                struct kernfs_elem_dir dir;
                struct kernfs_elem_symlink symlink;
                struct kernfs_elem_attr attr;
        };
        void *priv;
        union kernfs_node_id id;
        short unsigned int flags;
        umode_t mode;
        struct kernfs_iattrs *iattr;
};

One way of dealing with this would be to pass to tracee.bpf the current
running OS distro and kernel version and use branches to decide which
struct flavor to use.. but we cannot do that because of the dead branch
bug of v5.4 kernel.

RHEL8 4.18 kernel is also very different from vanilla 4.18 kernel.

Another example is the

struct pid_link {
        struct hlist_node node;
        struct pid *pid;
};

not existing in RHEL8 and the task_struct of RHEL8 having:

struct task_struct {
...
    struct pid *thread_pid;
...
}

Based on that, type the type/field checks are better than version checks
because most of the times it will do the right thing (if the type we
want exists we declare it as a flavor).