tracee-ebpf: turn CO-RE v4.18 and beyond compatible
Fixes: #1171

For the VANILLA v4.18 (and beyond) kernels, we need the following 3
commits backported, together with this patch, in order to have tracee
eBPF object loaded correctly.

commit 0962590e5533
Author: Daniel Borkmann <daniel@iogearbox.net>
Date:   Wed Oct 31 23:05:52 2018

    bpf: fix partial copy of map_ptr when dst is scalar

commit e4f8d81c738d
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date:   Mon Jul 9 21:48:54 2018

    cgroup/tracing: Move taking of spin lock out of trace event handlers

commit f7cf25b2026d
Author: Alexei Starovoitov <ast@kernel.org>
Date:   Sat Jun 15 12:12:17 2019 -0700

    bpf: track spill/fill of constants

There is no v4.18 stable tree, so there is no way to propose those
patches to upstream. Nevertheless, it is important to mention here so
someone needing v4.18 kernel is able to compile it with the 3 mentioned
patches - and possibly other dependencies - and this patch to tracee.

The v4.19 (stable tree: linux-4.19.y branch) kernel already contains the
needed fixes and doesn't need changes to have tracee CO-RE object
running.