Makefile.tracee: to have official 3 container flavors
This change makes it possible to have 3 containers generated:

1. tracee		# uses CO-RE by default
2. tracee-btfhub	# uses CO-RE by default and has embedded BTFs
3. tracee-nocore	# does not use CO-RE eBPF object (builds it)

The "tracee" container is the "slim" version and "tracee-nocore" is the
fat version of tracee (sort of speak).

If you want to execute "tracee-ebpf" instead of full tracee, those 3
flavors allow you to do so by exporting: TRACEE_EBPF_ONLY=1 environment
variable to docker. With that, the entrypoint will be changed
automatically to execute "tracee-ebpf" instead of "tracee".