tracee-ebpf: introduce event dependencies
Some events depend on other events to function properly.
Add a new field to event definition to declare those dependencies.
When an event is chosen by the user, set its dependencies essential.

This change will also help with the introduction of higher level events,
that are derived from other events received in userspace.