tracee-ebpf: fix container filters
Currently we don't trace existing container processes when a container
filter is set. This behavior is due to using the NewPidFilter when
a container filter is set. As we now use cgroups to identify
containers (and not by checking that pid != host_pid like we used to),
we can remove this extra filter to trace existing container processes
as expected by the filter.

Fix #1436