events_pipeline: add queueEvents to cache pipeline events
Currently tracee-rules has a throughput ratio of 1:4 (events consumed
versus events being produced by tracee-ebpf). This requires tracee-ebpf
to deal with this lag but, at the same time, the perf buffer events
consumption cannot be left behind (or we might lose important events
coming from kernel and miss signatures detections).

One way to solve this would be to increase the ChanEvents channel buffer
size so, if events being printed weren't consumed as fast as they were
being produced, the channel would cache them until they can be consumed,
releasing the pressure over the perf buffer.

Unfortunately, only increasing the ChanEvents buffer wasn't enough to
stop causing event losses. There is a need for a FIFO QUEUE in the
pipeline, so events are cached and released according to speed of the
end-of-pipeline consumption.

For now, this patch introduces in the pipeline a queue of up to 1
million events. This has proved to significantly reduce event losses
during host load bursts, avoid signature detection losses.

It has been tested in a 2GB environment (and higher): memory allocation
and de-allocation behaved well (~1GB consumption until the buffer was
full, when events started being lost again, something that is expected).

New tracee-ebpf requirement: 2GB of ram minimum.

Until we know good default cache sizes, based on performance benchmarks
of different host configurations, this option won't be enabled by
default. Idea is, in near future, to give end user the power of
controlling the size of this queue. It will be a game of how much memory
you are willing to spend versus how many event loss you are able to
accept.

Also, Future work might include queue watermarks advertisement for the
eBPF programs. This way, based on the queue usage, eBPF programs might
chose to send events (or not) based on their priorities. With that, the
pressure over the perf buffer would be reduced as the cache is consumed.