tracee-ebpf: perf-buffer-size set based on host size
An intermediate cache, in between the events production (tracee-ebpf)
and events consumption (tracee-rules), is needed, in order to allow
bursts of events not to cause event loss.

Fixes: #1456

This cache might be an internal (to tracee-ebpf) buffer, releasing the
pressure in the perf-buffer, OR it can be the perf-buffer itself (since
it is an anonymous pages mmap done by libbpf).

Having an internal asynchronous cache (and not perf-buffer only) is good
if we want to offload events to an external queue, for example. For now,
we don't need it, as we are not offloading the events. With that, we can
increase the perf-buffer size only and increase detection rates for
higher load bursts.

This commit sets default perf-buffer sizes based on the amount of memory
host has:

- host: up to 1  GB, perf-buffer-size = 256 MB
- host: up to 4  GB, perf-buffer-size = 512 MB
- host: up to 8  GB, perf-buffer-size = 1   GB
- host: up to 16 GB, perf-buffer size = 2   GB

For bigger hosts, the perf-buffer-size is 4 GB. The perf-buffer size
might be overriden by the cmd option:

--perf-buffer-size X

where X is the amount of memory pages reserved for the perf buffer (just
like it was before).