events_pipeline: add queueEvents to cache pipeline events
Under some circumstances, tracee-rules might be slower to consume
events than tracee-ebpf is capable of generating them. This requires
tracee-ebpf to deal with this possible lag, but, at the same time,
perf-buffer consumption can't be left behind (or important events coming
from the kernel might be loss, causing detection misses).

There are 3 variables connected to this issue:

1) perf buffer could be increased to hold very big amount of memory
   pages: The problem with this approach is that the requested space,
   to perf-buffer, through libbpf, has to be contiguous and it is almost
   impossible to get very big contiguous allocations through mmap after
   a node is running for some time.

2) raising the events channel buffer to hold a very big amount of
   events: The problem with this approach is that the overhead of
   dealing with that amount of buffers, in a golang channel, causes
   event losses as well. It means this is not enough to relief the
   pressure from kernel events into perf-buffer.

3) create an internal, to tracee-ebpf, buffer based on the node size.

This commit introduces (3): A generic interface for caching events in
the pipeline.

For now, there is a single interface implementation: an in-memory
caching mechanism with cmdline settings for cache type and possible
cache options. Needed tests were added to guarantee cmdline options are
always right.

In a near future, we may add new caching backends, such as offloading
events to filesystem, to a broken, etc.