tracee-ebpf: add events handle config
It is required to differentiate between different events configurations.
When an event is chosen, its bpf programs need to be loaded.
Some bpf programs might be used by different events, so it is required
to know which events should be submitted to userspace, and which
shouldn't.
In addition, not every event that is submitted to userspace needs to be
emited to the user. For example, when a derived event such as
container_create is chosen, the cgroup_mkdir event should be submitted
to userspace, however, if the user didn't choose the cgroup_mkdir event,
it shuoldn't be emitted as part of the output.

This commit dynamically creates the required configuration for each event
according to the chosen events and their dependencies.