pkg/events/derive: created new derive package
Create a derive package for the derived events implementations of
tracee-ebpf.
This way, there is a generic API for creating new derived events,
and derived events implementations can be created in different files.
The new API and division to files should make derived events implementations
more manageable and more testable.