tracee-ebpf: flatten event probe dependencies
Currently, in order to define which probes are required for an event to
function correctly, we either define the probes explicitly or use an
event dependency, which will load and attach the required probes.
This other method of defining the probes implicitly should be used only
when the event is a real dependency, and not just as a mechanism to load
and attach the other event's bpf program.

This commit explicitly defines which probes are required for an event to
function correctly, and always submit an event to userspace when it is
defined as a dependency of another event. In addition, it sets tracee to
always load all the bpf programs that exists on the BPF object file.
This is required to break the coupling between an event and the probes
used by it. It will also be required to support dynamically changing
events during runtime as described in [1]. Another future PR might not
load all the BPF probes if it will detect that not all the required BPF
dependencies are met (e.g. LSM-BPF is not supported by the running
kernel).