move enrichment logic to the event pipeline
This is done in order to ensure container enrichment queries
will not block the event pipeline.

Adds a stage where events are enriched with container runtime data.
Events will be enriched in the decode state if the data was
previously queried.
In that case, the pipeline stage will be skipped.