tracee-ebpf: use cached should_trace
To avoid performing scope caclculations (should_trace) on every event,
cache the last result of should_trace() in a new per-task map.

In addition, add task's start_time into the event context of every event
so we can identify a specific task even when a pid is being reused (due
to wrapping).