k8s deployment: increase needed limits because of caching
When caching was added to entrypoint the limits were not increased from
the deployment recipes, making tracee to fail with OOM in real
environments.

The default memory consumption shall be in between 768M and 1GB of ram
by default. That might be changed in a near future when performance
increases and less caching is needed.

For now, instead of limitting resources by default, we comment the
limits out with important observations regarding performance.