tracee-ebpf: mitigating missing caps for kprobes
Some distributions need CAP_SYS_ADMIN to dispatch kprobes, even
though they shouldn't need it according to kernel documentation.
This PR make CAP_SYS_ADMIN mandatory, with the hope to find the reason
for the bug in the future so CAPS_SYS_ADMIN
won't be needed anymore (in new kernels).