enrich: fix enrichment queue locks (#1877)
* enrich: read queue once before actions

This commit includes two fixes:
1. In the enrichment receiving part of this pipeline stage,
the relevant queue is read with a mutex acquire at multiple points.
To avoid unnecessary points of failure with that I've moved this
queue acquisition to the top of the loop.
2. In the enrich sending routine, the enrich is sent to a queue while
a lock is acquired. If the queue is full the send will be stuck and
the lock will not be unlocked. The queue is now read before sending the
enrichment so the lock will always be available.

This commit also increases the size of the enrichment queue buffers from
1000 to 2500.

This commit also include:

* enrich: fix send on closed channel

On some edge cases an event could be sent to an enrich queue which
was closed. Now, a map of stale queues exists to hold these queues so
the sending routine can close them.