containers: mount cpuset on cgroupsv1
When cgroupsV1 is used, the cgroup controller is located at the host
/sys/fs/cgroup/cpuset. However, this fs is duplicated to the container's
filesystem and used instead of it when running tracee-ebpf
from a container. As a result, a different root cgroupId is given,
and existing cgroups are not detected for population on startup.

This fix makes tracee create a local mountpoint for the host's cpuset
if it detects usage of cgroupsV1.