events_enrich: a new atomic approach with single lock
Producer:

Each cgroupId gets its own channel of events. Events are enqueued and only
de-dequed once the cgroupId was enriched OR has failed enrichment.

[cgroupId #1]: queue of trace events for container described by cgroupId #1
[cgroupId #2]: queue of trace events for container described by cgroupId #2
[cgroupId #3]: queue of trace events for container described by cgroupId #3
...

If the cgroupId channel does not exist it is created and an enrichment phase
is triggered for that cgroupId. If the cgroup channel is not used after a
specific period, it is removed.

Consumer:

In order to de-queue an event from its queue there is a scheduled operation
to each queued event.

queueReady: it is a simple scheduler of events using their cgroupId as index

[cgroupId #1][cgroupId #2][cgroupId #3][#cgroupId #2][cgroupId #3]...

One event is de-queued from queueReady only if its respective cgroupId has
been enriched OR has failed enrichment.

- - - -

Observations:

* With this model, the pipeline will only block when one of these
  channels is full. In this case, pipeline will be blocked until this
  channel's cgroupId is enriched and its enqueued events are de-queued.

* A good way to test this is to run the following:

  sudo ./dist/tracee-ebpf \
         --cache cache-type=mem \
         --cache mem-cache-size=1024 \
         --containers \
         -o format:json \
         -o option:parse-arguments \
         -o option:detect-syscall \
         -trace container \
         -trace event=openat \
         --crs docker:/var/run/docker.sock \
         | jq -c '. | {containerName}'

  during a stress test for container creations/deletions.

Commit changelog:

* Fix: execute routine only when needed
* Nest the type and consts inside the function to clean the namespace
* Add map cleanups on cgroup_rmdir
* Move enrichment to sending goroutine
* Remove host queues - send them immediately out
* Protect map accesses with bLock
* Checks if map keys do exist
* Fix atomic Toggle (set it to Store)

Signed-off-by: Rafael David Tinoco <rafaeldtinoco@gmail.com>
Signed-off-by: Nadav Strahilevitz <nadav.strahilevitz@aquasec.com>