enrich: fix lock in cleanup logic (cgroup_rmdir)
Similar bug to the previous version of enrichment:
When cleaning the queue one last time on a cgroup_rmdir event
the range loop would run forever because the channel isn't closed yet.
As such an explicit length check is required before and in the loop.

In addition, the channel is now closed before being deleted.