tracee: events: docs: fd arg enrichment
Introduce 'parse-arguments-fds' option that outputs a fd value in
the "fd=filepath" format.

Alert: when parse-arguments-fds is set, fd JSON's output value is
changed from integer to string. It also forces 'parse-arguments'.

'ParseArgsFDs()' is looking for events which have an argument named
"fd", disregarding the syscall ID. It can be improved to accomodate
syscalls with multiple fds and arguments with different naming.

Documentation is updated.