printer: add benchmarks
These benchmarks show the potential of protobuf encoding.

5 benchmarks are included:
1-3 are full printer benchmarks, 4 and 5 benchmark the marshaling alone.
Results 1-3 indicate currently that the protobuf printer is the least
efficient of the three.
However the rest of the benchmarks show the potential of protobuf
marshaling, when removing the protobuf wrapping from the equation, it is
clearly more efficient than gob as well.