tracee.bpf.c: optimize sys_enter and sys_exit
Optimize these hot path probes by moving the logic of event init and
submit to run conditionally in a tail call.
Only the following logic will run if no event submission is needed:
1. Storing task's current syscall
2. Storing syscalls arguments in task

further optimizations are done in the task_info intialization logic
which was moved to it's own function.