ebpf: add get_task_syscall_id helper
This is used in events supporting the detect-syscall option, in order to
decouple them from the syscall_data_t struct tracking mechanism.
This decoupling is required for these events to function in the future
without calling sys_enter and sys_exit logic for every syscall.
security_file_open is intentionally left as is, because it will have
syscall tail call dependencies in the near future.