pkg/events: fix socket option parse args edge cases
libbpf go currently parse the socket option argument in some cases to
include "or" in the middle between to consts names.
The reason is that the kernel has 2 consts with the same name.
According to the context we know which constant is relevant.
This PR fix the parsed option name to be the one we expect it to be.