printer: make sure output and errfile are always truncated
Output and error files should be always truncated.

It is imperative that either the output file is truncated OR that the
cursor positioning (to EOF) is done before any write to the file
happens. If not, then we might have situations such as:

$ cat ebpf_events.json | wc -l
58010
$ cat ebpf_events.json | wc -l
58010
$ cat ebpf_events.json | wc -l
58011
$ cat ebpf_events.json | wc -l
58002
$ cat ebpf_events.json | wc -l
57989
$ cat ebpf_events.json | wc -l
57978

when the output file is truncated while being written to.