tracee: remove invalid events from tailcalls
On some architecture the default tail calls may include events that do
not exist, for example in ARM64 the "open" syscall is missing but
security_file_open requires it as a traced syscall in "sys_enter_init_tails".

As such when extracting tail calls during initialization we remove the
invalid event ids from tail calls.
In addition, when getting a direct tracing request for a syscall event
which does not exist, we skip adding it's index to the syscall tracing
tail call maps.