Run integration test triggers in own PID
This is a change to the integration tests so that the commands
which are meant to trigger events that the tests look for, are
actually picked up by tracee. In order for this to happen,
the events can't be the same PID as tracee. This is accomplished
using syscall.ForkExec() which runs functions in a shell script.

Signed-off-by: grantseltzer <grantseltzer@gmail.com>