Change BPF object parsing logic to support multiple objects
This changes the fields in Tracee's config and central
data structure so that there are multiple possible BPF
compiled BPF objects with their own paths and contents.

This PR should not change any logic or behavior, but
instead set tracee up for modular BPF objects where
current 'main' BPF object is replaced by a series of
modules such as for individual programs and maps.

Signed-off-by: grantseltzer <grantseltzer@gmail.com>