ebpf: add proc_info_map
Currently, when the user specifies pid=new filter, tracee actually
traces new threads, even if those are of existing processes. This
behavior is confusing, and doesn't fit the name of the filter.

Introduce a new process info bpf map which will keep track of processes
in the system. In addition to solving the above problem, this will also
allow us to add new fields in the future that are process scoped, e.g.
process binary name.