ebpf: introduce sched_process_free event
This new event is attaching the sched_process_free tracepoint and can be
used for the deletion of per process maps. This is done by comparing pid
to tgid, and performing the deletion only when tgid exits. Unlike
sched_process_free, here we can be sure that process exited since if the
group leader exited first, it will become a zombie until all other
threads in the group exited. Only then sched_process_free is called for
the group leader.