pkg/events: change symbols_loaded to use logger
Previously we returned an error upon failure in the event, even though
the error in the event is expected and spam the user.
Now that we have the logger, we can give warning and debug messages
instead, avoid spam of error logs but allowing developers to monitor
errors in the event.