pkg/events: add API to derive multiple events from single function
Until now, we had a function to support derivation of a single event
from a single base event.
However, sometimes a single base event can trigger multiple instances
of a single derived event (aka, one derive function that should derive
multiple events in a single call).
We add support for it, and unite the two APIs to use the same code.