filters: expand binary filter to support mount ns
Currently the binary filter can not differentiate between binaries that
have the same path but reside in different mount namespaces. Having the
mount namespace is required to identify a specific binary to filter by.

Enable the user to provide the mount namespace in addition to the binary
path and keep support of providing a binary without specifying a mount
namespace.

To filter by path, the following syntax is used::
    --trace binary=/some/binary/path

To filter by path and a specific mount namespace (e.g. 4026532448), use:
    --trace binary=4026532448:/some/binary/path

Specifying the host mount namespace can be done using the "host" alias:
    --trace binary=host:/some/binary/path