ebpf: fix process tree filter
BPF code used to check if a process should be filtered by the process
tree filter according to the namespace pid and not the host pid.
Doing this, same pids in different namespaces will be filtered and not
only a single process hierarchy. Fix this by checking host_pid instead.