events_pipeline: run filters for derived events
When the cmdline has arg filters for derived events, for example, they
are never filtered (Equal, NotEqual, Contains, NotContains) as their
Event ID comes from the functions DeriveSingleEvent & DeriveMultiple
Events, and not from the perfbuffer event, like regular events.

Their filtering stage should be done in a later stage in the pipeline,
when derivatives (derived events) are obtained, and not in the
decodeEvents phase. This patch fixes the issue by running all filters
for the derived event being processed.

Fixes: #2462