filters: if arg is missing do not process it
When filtering an event for an argument that might not exist (because
the eBPF code might not send it), the events missing the argument
being filtered were being sent.

One exampe is when openat filters for openat.args.pathname of
/etc/passwd. Some openat events missing pathname might be sent out
through printer.

Do not process the event when an argument being filtered for does not
exist in the event.

Fixes: #2487