pipeline: change event processing to be map based
The event processing part of the pipeline's mechanism is changed from a
hard coded switch statement to use a map which registers for each event
id it's processing functions.
By using this mechanism we can expose a function to register such a
processor programatically, which will be useful later for future tracee
event plugins.