refactor: use a map to store event data
Currently event data is stored on the stack using a struct named
event_data_t. This has the drawback of consuming 120 bytes
(sizeof(event_data_t)) out of the available 512 bytes before we even
begin considering the required logic of the event. By moving this struct
into a map, not only we save this space on the stack, but we can then
also add a buffer to save the event's arguments instead of having a
dedicated submit buffer. This allows us to place the buffer right after
the context of the event, and send it to userspace without having to do
an extra copy of the context to the submit buffer.

In addition to the above benefits, this also sets the ground to split
our bpf programs into several programs linked by tail calls and keeping
the event's state on the event data map.