integration: add tests for binary filter
First test confirms that `docker run` commands output with a filter on
dockerd's binary path will include events with it's own PID, which was
previously impossible as the original process would not exist in the
process info map on startup.

The second test is a simple integration test for the binary filter with
separate scopes.