types/trace: add kernel read type
In the SecurityKernelReadFile and SecurityPostReadFile events the
"type" argument is used to indicate the object read by the kernel.
However this type is a volatile enum in the kernel that may change
by version.
As such it is normalized as part of the parse-arguments output option.

In order to remove the need for parse-arguments in signature writing,
this argument needs to be normalized to some enum value in tracee,
which can later be string translated in parse-arguments like before.
This type is meant to be used for that.