events: normalize kernel read type
Use the new KernelReadType (commit 188803b23e14af0e41) in order to
normalize the "type" argument of SecurityKernelReadFile and
SecurityPostReadFile.
Converting the argument from int to the enum is done in the processing
stage instead of the parsing stage, so signatures can use the constant
enum values.

String values of the enum can still be used by providing the
parse-arguments option in the output flag.