mscope: ebpf: check event submit by scope with map
When using multiple scopes, probes may be attached due to a single scope
but it's events not being required at all by other scopes.

As such, checking if the event is required, by any of the matched
scopes, should be done as part of checking an event submit.

As part of this patch, should_submit checks were added in most kprobe
programs.

They were previously unnecessary because they would only be attached
as part of the single scope provided. With the addition of multiple
scopes it is not guarenteed that probe attaching is equivalent to a
submit check (if there is a one to one relation between probe and
event).