pipeline: copy the base event before deriving
This is needed because later modification of the event (in particular
of the matched scopes) can affect the derivation and later pipeline
logic acting on the derived event.

In particular this effect was noticed in container_create since it is
often used without cgroup_mkdir. As such when adjusting scopes before
emitting (see sinkEvents) the MatchedScopes would be zeroed, affecting
the MatchedScopes passed to the derived container_create event.