k8s: add kustomize templates
This will be used to generate all k8s static files.
It allows one single file to be the base, and customizations
are based on kustomize patches. For example, the same base
will generate tracee.yaml, tracee-falcosidekick.yaml, and
tracee-postee.yaml, requiring change only on the base file for
any change.